:recycle: [Refactor] dto/response 디렉토리 생성
