+++ src/main/java/com/gamegoo/gamegoo_v2/chat/dto/ChatResponseFactory.java
@@ -2,6 +2,11 @@ package com.gamegoo.gamegoo_v2.chat.dto;
 
 import com.gamegoo.gamegoo_v2.block.service.BlockService;
 import com.gamegoo.gamegoo_v2.chat.domain.Chat;
+import com.gamegoo.gamegoo_v2.chat.dto.response.ChatMessageListResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.ChatMessageResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.EnterChatroomResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.SystemFlagResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.SystemMessageResponse;
 import com.gamegoo.gamegoo_v2.friend.service.FriendService;
 import com.gamegoo.gamegoo_v2.member.domain.Member;
 import lombok.RequiredArgsConstructor;
