+++ src/main/java/com/gamegoo/gamegoo_v2/chat/dto/response/ChatMessageListResponse.java
@@ -1,4 +1,4 @@
-package com.gamegoo.gamegoo_v2.chat.dto;
+package com.gamegoo.gamegoo_v2.chat.dto.response;
 
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
