+++ src/main/java/com/gamegoo/gamegoo_v2/chat/dto/response/SystemFlagResponse.java
@@ -1,4 +1,4 @@
-package com.gamegoo.gamegoo_v2.chat.dto;
+package com.gamegoo.gamegoo_v2.chat.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
