+++ src/main/java/com/gamegoo/gamegoo_v2/chat/service/ChatFacadeService.java
@@ -3,9 +3,9 @@ package com.gamegoo.gamegoo_v2.chat.service;
 import com.gamegoo.gamegoo_v2.chat.domain.Chat;
 import com.gamegoo.gamegoo_v2.chat.domain.Chatroom;
 import com.gamegoo.gamegoo_v2.chat.domain.MemberChatroom;
-import com.gamegoo.gamegoo_v2.chat.dto.ChatMessageListResponse;
 import com.gamegoo.gamegoo_v2.chat.dto.ChatResponseFactory;
-import com.gamegoo.gamegoo_v2.chat.dto.EnterChatroomResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.ChatMessageListResponse;
+import com.gamegoo.gamegoo_v2.chat.dto.response.EnterChatroomResponse;
 import com.gamegoo.gamegoo_v2.common.validator.BlockValidator;
 import com.gamegoo.gamegoo_v2.common.validator.MemberValidator;
 import com.gamegoo.gamegoo_v2.exception.ChatException;
